first proposals of examples for the widget
Once the PR that allows to select a language and an example type is merged, this PR offers several examples in several languages of prompts.

I tried to do my best to be compatible with the yaml format but it's not impossible that some adjustments have to be made for the line breaks and the special characters such as the quotation mark.